refactor(fieldmap): colocate palette calculations inside PlotGrid
